Wow.  It... it WORKS!  It's awesome!  That's right, folks, we've got point-n-click moving, and it ROCKS!  Also, pressing 'T' toggles transport mode.
Not only that, but the code in MovePanel has been heavily refactored and is a whole lot nicer now.  Still needs to be organized appropriately, but that'll come later.

I tweaked the click/double-click timeout down to 350ms (default was 500, which was just too slow).

Finally, added a "YesNoForm".  May not get around to using it, but it's in there.